requirements.txt
certifi==2020.6.20
configparser==5.0.0
minio==6.0.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
urllib3==1.25.10
